Create Testrun and parse_junit_xml takes in filename
The Testrun class replaces the dict we would previously return from the
parse_junit_xml function.

Testrun contains 4 values: name, duration, outcome and testsuite.

The first 3 values were previously contained in the dictionary,
the new value is testsuite which is now read from the xml.

parse_junit_xml also now takes the filename of the file to
parse as its argument instead of a string of the contents of
the file.